Modificacion: permitir que ingrese una cantidad mas grande en las proyecciones cinco anios optimista, conservador y pesimista

+++ database/migrations/2024_09_17_172657_create_ingresos_cinco_anios_pesimistas.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('Id_estudio_financiero')->constrained('estudio_financieros')->onDelete('cascade');
             $table->foreignId('Id_ingresos')->constrained('ingresos')->onDelete('cascade');
             $table->integer('anio');
-            $table->float('monto_pesimista');
+            $table->decimal('monto_pesimista',12,2);
             $table->timestamps();
         });
     }
