+++ database/migrations/2024_10_02_193824_create_ingresos_cinco_anios_conservadors_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('Id_estudio_financiero')->constrained('estudio_financieros','id','fk_ingresos_estudio_financiero')->onDelete('cascade');
             $table->foreignId('Id_ingresos')->constrained('ingresos','id','fk_ingreso_pertenece')->onDelete('cascade');
             $table->integer('anio');
-            $table->float('monto_conservador');
+            $table->decimal('monto_conservador',12,2);
             $table->timestamps();
         });
     }
