+++ database/migrations/2024_10_03_211430_create_ingresos_cinco_anios_optimistas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('Id_estudio_financiero')->constrained('estudio_financieros','id','fk_ingresosOptimistas_estudio_financiero')->onDelete('cascade');
             $table->foreignId('Id_ingresos')->constrained('ingresos','id','fk_ingreso_optimista_pertenece')->onDelete('cascade');
             $table->integer('anio');
-            $table->float('monto_optimista');
+            $table->decimal('monto_optimista',12,2);
             $table->timestamps();
         });
     }
